requirements.txt
numpy==1.24.2
openai==0.27.4
pandas==2.0.0
pdfminer.six
python-dotenv==1.0.0
tiktoken==0.3.3
wget==3.2
nltk
habanero
bs4
diskcache